container refactored to a hook
